refactor: keep map_children_mut private to the optimizer crate
`map_children_mut` was added as a public method on `LogicalPlan` in
`datafusion-expr` only because the optimizer, in a different crate,
needed to call it. But the optimizer is its sole consumer, and the
`Arc::make_mut` in-place trick does not generalize to the other tree
types (`Expr` children are `Box`ed; `PhysicalExpr`/`ExecutionPlan`
children are `Arc<dyn _>`, which `Arc::make_mut` cannot handle), so
committing to it as public API is not warranted.

Move it into `optimizer.rs` as a private free function next to
`rewrite_plan_in_place`, its only caller. `datafusion-expr` is now
untouched by this PR except for the `map_subqueries` short-circuit, and
the optimizer adds no public API. If `TreeNode` ever grows an in-place
traversal this logic can move there with no breaking change.

No behavior change; the 713 datafusion-optimizer tests pass unchanged.
